+++ src/steps/active-directory/converters.ts
@@ -32,7 +32,7 @@ export function createAccountEntity(instance: IntegrationInstance): Entity {
 export function createAccountEntityWithOrganization(
   instance: IntegrationInstance,
   organization: Organization,
-  extraData: {
+  intuneConfig: {
     mobileDeviceManagementAuthority?: string;
     subscriptionState?: string;
     intuneAccountID?: string;
@@ -48,20 +48,24 @@ export function createAccountEntityWithOrganization(
 
   return createIntegrationEntity({
     entityData: {
-      source: { ...organization, ...extraData },
+      source: {
+        organization,
+        intuneConfig,
+      },
       assign: {
         _class: entities.ACCOUNT._class,
         _key: `${entities.ACCOUNT._type}-${instance.id}`,
         _type: entities.ACCOUNT._type,
+        id: organization.id,
         name: organization.displayName,
         displayName: instance.name,
         organizationName: organization.displayName,
         defaultDomain,
         verifiedDomains,
-        intuneAccountId: extraData?.intuneAccountID,
+        intuneAccountId: intuneConfig?.intuneAccountID,
         mobileDeviceManagementAuthority:
-          extraData?.mobileDeviceManagementAuthority,
-        intuneSubscriptionState: extraData?.subscriptionState,
+          intuneConfig?.mobileDeviceManagementAuthority,
+        intuneSubscriptionState: intuneConfig?.subscriptionState,
       },
     },
   });